small change of db, tables

+++ DBServer/Controllers/CompaniesControler.cs
@@ -24,7 +24,7 @@ namespace DBServer.Controllers
 
         // GET api/values/5
         [HttpGet("{id}")]
-        public ActionResult<Company> Get(int id)
+        public ActionResult<Company> Get(long id)
         {
             return _context.Companies.Single(c => c.CompanyId == id);
         }
@@ -39,13 +39,13 @@ namespace DBServer.Controllers
 
         // PUT api/values/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] string value)
+        public void Put(long id, [FromBody] string value)
         {
         }
 
         // DELETE api/values/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public void Delete(long id)
         {
             var delCompany = _context.Companies.Single(c => c.CompanyId == id);
             _context.Companies.Remove(delCompany);
